Each "example" task now has its own directory.
