+++ CART/CARTLib/examples/OrganLabellingDemo/OrganLabellingDemo.py
@@ -1,9 +1,9 @@
 from typing import Optional
 
-from .core.TaskBaseClass import TaskBaseClass, DataUnitFactory
-from .core.DataUnitBase import DataUnitBase
+from ...core.TaskBaseClass import TaskBaseClass, DataUnitFactory
+from ...core.DataUnitBase import DataUnitBase
+from ...LayoutLogic import CaseIteratorLayoutLogic
 from .VolumeOnlyDataIO import VolumeOnlyDataUnit
-from .LayoutLogic import CaseIteratorLayoutLogic
 
 import ctk
 import qt
